API-48806: Remove Benefits Intake v2 API, Part 6 (`UploadSerializers`, Webhook Registrations and Events, Env Vars) (#24338)
* API-48806: Remove v2 UploadSerializer and shared serializer

* API-48806: Remove v2-specific UploadSubmission model code

* API-48806: Remove webhook registration code

* API-48806: Remove Benefits Intake v2 environment variables

* API-48806: Remove Benefits Intake v2 webhook-specific env vars

* API-48806: Comment out webhook registration require to resolve failing test

* API-48806: Resolve failing specs

* API-48806: Resolve rubocop offense

---------
